Implement class model (Abandon code generated by Umple)

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/library/dao/LibraryItemRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import ca.mcgill.ecse321.library.models.LibraryItem;
 
-public interface LibraryItemRepository extends CrudRepository<LibraryItem, Integer>{
+public interface LibraryItemRepository extends CrudRepository<LibraryItem, Long>{
 	
-	LibraryItem findLibraryItemByBarcode(int barcode);
+	LibraryItem findLibraryItemByBarcode(Long barcode);
 
 }
